feat: add JWT service and dependencies

+++ src/main/java/brenno/AgredadorInvestimentos/Controller/Dto/AccountStockResponseDto.java
@@ -1,4 +1,4 @@
 package brenno.AgredadorInvestimentos.Controller.Dto;
 
-public record AccountStockResponseDto(String stockId, int quantity, double total) {
+public record AccountStockResponseDto(String stockId, double quantity, double total) {
 }
